Create Toyota Corolla Used Car Validation
These are the result of fitting a certain predictive model to prices of used Toyota Corolla cars. The training set includes 600 cars and the validation set includes 400 cars. Results are displayed separately for the training and validation sets. We can see from the histogram and boxplot corresponding to the validation set that most errors are in the [−1000, 1000] range, with a few large positive (under-prediction) errors